+++ Vault-Coder/Changes.md
@@ -9,6 +9,7 @@
 - Added support for **Days** and **FiniteDuration** data types (both concrete and optional variants)
 - Added custom indexing support (see README for details)
 - The app can now write specific classes, packages or enumerations based on filtering applied on program use
+- Written many model access points now override the `filter(Condition)` function
 - Written Stored models now contain a property `access`, which points to their id-specific access point
 ### Bugfixes
 - Optional references didn't work previously - now fixed